Introducing peak intensity
It represents intensity of a peak. It is used to verify the current peak observed
is strong enough to be considered as true.
Here we get the maximum intensity of the current peak and compare it to the last 5 max
intensities of the peak observed.

It is computed as the sum of the last 3 vectors (3 beacause an increasing front is
3 successive positive vectors).
It replaces the maxVector used previously.